+++ src/at/sume/dm/model/residential_satisfaction/DesiredLivingSpace.java
@@ -19,9 +19,9 @@ public class DesiredLivingSpace extends ResidentialSatisfactionComponent {
 	 * @see at.sume.dm.model.residential_satisfaction.ResidentialSatisfactionComponent#calc(at.sume.dm.entities.HouseholdRow, at.sume.dm.entities.SpatialUnitRow, int)
 	 */
 	@Override
-	public long calc(HouseholdRow household, DwellingRow dwelling, SpatialUnitRow spatialUnit, int modelYear) {
+	public short calc(HouseholdRow household, DwellingRow dwelling, SpatialUnitRow spatialUnit, int modelYear) {
 		long currentLivingSpace = 0;
-		long result = 0;
+		short result = 0;
 		// TODO: add household-specific desiredLivingSpace modifier here
 		household.estimateDesiredLivingSpace();
 		long desiredLivingSpace = (household.getAspirationRegionLivingSpaceMin() + household.getAspirationRegionLivingSpaceMax()) / 2;
@@ -43,7 +43,7 @@ public class DesiredLivingSpace extends ResidentialSatisfactionComponent {
 			}
 		}
 		assert currentLivingSpace >= 0 : "Current living space < 0 (" + currentLivingSpace + ") for household " + household.getId();
-		result = Math.round(currentLivingSpace * 1000 / desiredLivingSpace);
+		result = (short) Math.round(currentLivingSpace * 1000 / desiredLivingSpace);
 		if (result > 1000)
 			return 1000;
 		return result;
